+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## next
 
+- rename image_classification
 - more examples
+- using a better product for product image classification test and example
+- add log4js
+- update translate service to 1810B (https://help.sap.com/viewer/2e6173bf645243bb9a88f0269250f3a2/1.0/en-US)
 
 ## 0.5.6